+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 0.0.50
 
+-   Email Templates are configurable from content API
 -   Take CSW connector distribution access url and format from distributor field under distributionInfo as well as from transfer options (previous.)
 -   Delete registry API /records/{recordId}/aspects endpoint
 -   Hook actors will report its status to its parent actor when changes